requirements.txt
aiohttp==3.10.11
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
cachetools==5.2.0
certifi==2024.7.4
charset-normalizer==2.1.0
cycler==0.11.0
discord.py==2.2.2
fonttools==4.43.0
frozenlist==1.3.0
google-auth==2.9.1
google-auth-oauthlib==0.5.2
gspread==5.4.0
gspread-dataframe==3.3.0
httplib2==0.20.4
idna==3.7
kiwisolver==1.4.4
matplotlib==3.5.2
multidict==6.0.2
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
pandas==1.4.3
Pillow==10.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.32.4
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
toml==0.10.2
urllib3==1.26.19
yarl==1.7.2
stripe==7.0.0
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0
